third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-language-override-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['font-language-override'] = "normal" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['font-language-override'] = "\\"KSW\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['font-language-override'] = "\\"APPH\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['font-language-override'] = "\\"ENG \\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['font-language-override'] = "\\"ksw\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['font-language-override'] = "\\"tr\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['font-language-override'] = "\\"en  \\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['font-language-override'] = "\\" en \\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['font-language-override'] = "\\"1 %\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
